pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pr_firm"
version = "0.1.0"
description = "A PocketFlow-based LLM application for brand-compliant marketing copy generation"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "pocketflow>=0.0.1",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0",
    "gradio>=4.31.0",
    "openai>=1.30.0",
    "requests",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy",
    "flake8",
    "black",
    "isort",
]

[project.scripts]
pr_firm = "pr_firm.cli:main_cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
pr_firm = ["data/**/*"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.black]
line-length = 79
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 79

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=pr_firm --cov-report=html --cov-report=xml"